Fix betty fails

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -3,7 +3,7 @@
 #include "function_pointers.h"
 
 /**
- * array_iterator - Executes a function given as 
+ * array_iterator - Executes a function given as
  * a parameter on each element of an array.
  *
  * @array: The array of integers numbers.
@@ -12,7 +12,7 @@
  **/
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	long unsigned int i;
+	unsigned int i;
 
 	if (array != NULL && action != NULL)
 	{
